Sanket Pandit/Day 4/requirements.txt
langchain
python-dotenv
langchain-groq
requests
